move irc to lib dir and circ to src dir
