+++ lib/irc/irc.c
