+++ lib/irc/irc.h
